AgentApp/requirements.txt
# Core
streamlit>=1.32.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0

# LangGraph / LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Tool calling utilities
langchain-openai>=0.2.0          # provides convert_to_openai_tool helper

# Web search (optional — only needed for general_question routing)
langchain-tavily>=0.1.0

# FastAPI fallback (optional — kept for app.py)
fastapi>=0.104.0
uvicorn>=0.23.2
